src/main/java/com/shawn/touchstone/chainOfReponsibility/handlers/Handler.java
package com.shawn.touchstone.chainOfReponsibility.handlers;

public abstract class Handler {
    protected Handler successor = null;

    public void setSuccessor(Handler successor) {
        this.successor = successor;
    }

    public Handler getSuccessor() {
        return successor;
    }

    public final void handle() {
        boolean handled = doHandle();
        if (successor != null && !handled) {
            successor.handle();
        }
    }

    public abstract boolean doHandle();
}
